+++ lab_2/constants.py
@@ -1,0 +1,1 @@
+pi = [0.2148, 0.3672, 0.2305, 0.1875]
