--- lab_2/utils/constants.py
@@ -1,4 +1,0 @@
-pi0 = 0.2148
-pi1 = 0.3672
-pi2 = 0.2305
-pi3 = 0.1875
